requirements.txt
flask == 3.1.1
flask-cors == 6.0.0
google-generativeai == 0.8.5
gspread == 6.2.1
python-dotenv == 1.1.0
gunicorn
